requirements.txt
aiohttp==0.20.2
aslack==0.11.0
chardet==3.0.4
python-dotenv==0.8.2
